% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ui.R
\name{aside_section}
\alias{aside_section}
\title{Create a section of a step's aside}
\usage{
aside_section(title, ...)
}
\arguments{
\item{title}{Section title tag, shown as a small uppercase eyebrow.}

\item{...}{Section content.}
}
\value{
A div.
}
\description{
Create a section of a step's aside
}
\keyword{internal}
