% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{background_jobs}
\alias{background_jobs}
\title{Whether slow jobs run in the background}
\usage{
background_jobs()
}
\value{
A logical scalar: \code{TRUE} when they run on daemons.
}
\description{
Whether slow jobs run in the background
}
\keyword{internal}
