% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_threshold_percent}
\alias{calculate_threshold_percent}
\title{Calculate threshold percent from concentration with rounding}
\usage{
calculate_threshold_percent(fit, conc, digits = 3)
}
\arguments{
\item{fit}{A fitdists object}

\item{conc}{Numeric concentration value}

\item{digits}{Number of significant digits for rounding (default: 3)}
}
\value{
Numeric percent value (0-100 scale)
}
\description{
Calculate threshold percent from concentration with rounding
}
\keyword{internal}
