% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clean_nboot}
\alias{clean_nboot}
\title{Clean bootstrap sample count from formatted string}
\usage{
clean_nboot(x)
}
\arguments{
\item{x}{Character string representing a number with separators}
}
\value{
Integer value with separators removed
}
\description{
Clean bootstrap sample count from formatted string
}
\keyword{internal}
