% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clean_ssd_data}
\alias{clean_ssd_data}
\title{Clean uploaded data}
\usage{
clean_ssd_data(data)
}
\arguments{
\item{data}{A data frame or tibble}
}
\value{
A cleaned tibble
}
\description{
Clean uploaded data
}
