% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ui.R
\name{dt_weight_color_bar}
\alias{dt_weight_color_bar}
\title{Add weight column color bar to a DT datatable}
\usage{
dt_weight_color_bar(dt, data, trans)
}
\arguments{
\item{dt}{A DT datatable object}

\item{data}{The data frame used to create the datatable}

\item{trans}{Translations reactive value}
}
\value{
The DT datatable with color bar formatting on the weight column
}
\description{
Add weight column color bar to a DT datatable
}
\keyword{internal}
