% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{estimate_hc}
\alias{estimate_hc}
\title{Estimate hazard concentration}
\usage{
estimate_hc(x, percent)
}
\arguments{
\item{x}{A fitdists object from ssd_fit_bcanz()}

\item{percent}{Numeric percent of species affected (0-100 scale)}
}
\value{
Numeric concentration estimate
}
\description{
Estimate hazard concentration
}
\keyword{internal}
