% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{estimate_hp}
\alias{estimate_hp}
\title{Estimate hazard percent}
\usage{
estimate_hp(x, conc)
}
\arguments{
\item{x}{A fitdists object from ssd_fit_bcanz()}

\item{conc}{Numeric concentration value}
}
\value{
Numeric proportion estimate (0-1 scale)
}
\description{
Estimate hazard percent
}
\keyword{internal}
