% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{estimate_time}
\alias{estimate_time}
\title{Estimate bootstrap computation time}
\usage{
estimate_time(nboot, lang)
}
\arguments{
\item{nboot}{Integer number of bootstrap samples}

\item{lang}{Character string language code: "english", "french", or "spanish"}
}
\value{
Character string with formatted time estimate
}
\description{
Estimate bootstrap computation time
}
\keyword{internal}
