% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{format_r_code}
\alias{format_r_code}
\title{Format R code}
\usage{
format_r_code(code_lines)
}
\arguments{
\item{code_lines}{Character vector of R code lines}
}
\value{
Single character string of the code
}
\description{
Joins the lines of code, which are written in tidyverse style, and quotes
their strings with single quotes.
}
\keyword{internal}
