% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ui.R
\name{gof_header_tooltips}
\alias{gof_header_tooltips}
\title{Build GOF-style header tooltips named vector}
\usage{
gof_header_tooltips(trans, lang = "english")
}
\arguments{
\item{trans}{Translations reactive value}

\item{lang}{Character string language code ("english", "french", "spanish")}
}
\value{
Named character vector mapping column names to tooltip descriptions
}
\description{
Build GOF-style header tooltips named vector
}
\keyword{internal}
