% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-plots.R
\name{gp_xbreaks}
\alias{gp_xbreaks}
\title{Extract and format x-axis breaks from ggplot}
\usage{
gp_xbreaks(gp)
}
\arguments{
\item{gp}{A ggplot2 object}
}
\value{
Numeric vector of sorted, rounded x-axis break values
}
\description{
Extract and format x-axis breaks from ggplot
}
\keyword{internal}
