% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{guess_conc}
\alias{guess_conc}
\title{Guess concentration column name}
\usage{
guess_conc(name, data = NULL)
}
\arguments{
\item{name}{Character vector of column names}

\item{data}{Optional data frame to identify numeric columns}
}
\value{
Character string of matched column name, or NA if no match
}
\description{
Guess concentration column name
}
\keyword{internal}
