% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{guess_sp}
\alias{guess_sp}
\title{Guess species column name}
\usage{
guess_sp(name)
}
\arguments{
\item{name}{Character vector of column names}
}
\value{
Character string of matched column name, or NA if no match
}
\description{
Guess species column name
}
\keyword{internal}
