% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{has_finite_concentration}
\alias{has_finite_concentration}
\title{Check if concentration values are all finite}
\usage{
has_finite_concentration(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
TRUE if all finite, FALSE otherwise
}
\description{
Check if concentration values are all finite
}
\keyword{internal}
