% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{has_min_concentration}
\alias{has_min_concentration}
\title{Check if there are at least 6 concentration values}
\usage{
has_min_concentration(x)
}
\arguments{
\item{x}{A vector}
}
\value{
TRUE if at least 6 values, FALSE otherwise
}
\description{
Check if there are at least 6 concentration values
}
\keyword{internal}
