% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{has_not_all_identical}
\alias{has_not_all_identical}
\title{Check if concentration values are not all identical}
\usage{
has_not_all_identical(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
TRUE if values are not all identical, FALSE otherwise
}
\description{
Check if concentration values are not all identical
}
\keyword{internal}
