% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{has_numeric_concentration}
\alias{has_numeric_concentration}
\title{Check if concentration values are numeric}
\usage{
has_numeric_concentration(x)
}
\arguments{
\item{x}{A vector}
}
\value{
TRUE if numeric, FALSE otherwise
}
\description{
Check if concentration values are numeric
}
\keyword{internal}
