% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{has_percent}
\alias{has_percent}
\title{Whether predictions include a percent affected}
\usage{
has_percent(pred, percent)
}
\arguments{
\item{pred}{Model-averaged hazard concentrations.}

\item{percent}{Numeric scalar percent of species affected.}
}
\value{
A flag.
}
\description{
Whether predictions include a percent affected
}
\keyword{internal}
