% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{has_positive_concentration}
\alias{has_positive_concentration}
\title{Check if concentration values are all positive}
\usage{
has_positive_concentration(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
TRUE if all positive, FALSE otherwise
}
\description{
Check if concentration values are all positive
}
\keyword{internal}
