% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{hint}
\alias{hint}
\title{Create grey hint text}
\usage{
hint(x)
}
\arguments{
\item{x}{Character string hint text}
}
\value{
HTML object with grey font color
}
\description{
Create grey hint text
}
\keyword{internal}
