% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{inline}
\alias{inline}
\title{Create inline-block div wrapper}
\usage{
inline(x)
}
\arguments{
\item{x}{Shiny UI element(s) to wrap}
}
\value{
tags$div with inline-block styling
}
\description{
Create inline-block div wrapper
}
\keyword{internal}
