% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{label_mandatory}
\alias{label_mandatory}
\title{Add mandatory field indicator to label}
\usage{
label_mandatory(label)
}
\arguments{
\item{label}{Character string or tag for the label}
}
\value{
tagList with label and asterisk span
}
\description{
Add mandatory field indicator to label
}
\keyword{internal}
