% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/about.R
\name{methods_page}
\alias{methods_page}
\title{Create the Methods page}
\usage{
methods_page(html)
}
\arguments{
\item{html}{Character string rendered About HTML for the current language.}
}
\value{
The page's content.
}
\description{
Create the Methods page
}
\keyword{internal}
