% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ui.R
\name{observe_step_button}
\alias{observe_step_button}
\title{Open a step when its button is clicked}
\usage{
observe_step_button(input, id, step, session = getDefaultReactiveDomain())
}
\arguments{
\item{input}{The module's input.}

\item{id}{Character string input ID of the button.}

\item{step}{Character string value of the step's tab in \code{main_nav}.}

\item{session}{The module's session.}
}
\value{
An observer.
}
\description{
Open a step when its button is clicked
}
\keyword{internal}
