% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{paste_js}
\alias{paste_js}
\title{Create JavaScript conditional for Shiny output}
\usage{
paste_js(x, ns)
}
\arguments{
\item{x}{Character string output name}

\item{ns}{Shiny namespace function}
}
\value{
Character string with JavaScript expression
}
\description{
Create JavaScript conditional for Shiny output
}
\keyword{internal}
