% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{remove_blank_headers}
\alias{remove_blank_headers}
\title{Remove Excel/Numbers blank header columns (X1, X2, etc.)}
\usage{
remove_blank_headers(data)
}
\arguments{
\item{data}{A data frame or tibble}
}
\value{
A tibble with X1, X2, ... X200 columns removed
}
\description{
Remove Excel/Numbers blank header columns (X1, X2, etc.)
}
\keyword{internal}
