% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{remove_empty_columns}
\alias{remove_empty_columns}
\title{Remove completely empty columns}
\usage{
remove_empty_columns(data)
}
\arguments{
\item{data}{A data frame or tibble}
}
\value{
A tibble with empty columns removed
}
\description{
Remove completely empty columns
}
\keyword{internal}
