% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgtemplate-package.R
\docType{package}
\name{shinyssdtools-package}
\alias{shinyssdtools}
\alias{shinyssdtools-package}
\title{shinyssdtools: ssdtools Shiny App}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Shiny application for fitting species sensitivity distributions using ssdtools.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bcgov/shinyssdtools}
  \item Report bugs at \url{https://github.com/bcgov/shinyssdtools/issues}
}

}
\author{
\strong{Maintainer}: Seb Dalgarno \email{seb@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-3658-4517}{ORCID})

Authors:
\itemize{
  \item Seb Dalgarno \email{seb@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-3658-4517}{ORCID})
}

Other contributors:
\itemize{
  \item Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID}) [contributor]
  \item Ayla Pearson \email{ayla@poissonconsulting.ca} (\href{https://orcid.org/0000-0001-7388-1222}{ORCID}) [contributor]
}

}
\keyword{internal}
