% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{silent_plot}
\alias{silent_plot}
\title{Print plot silently}
\usage{
silent_plot(x)
}
\arguments{
\item{x}{A plot object (typically ggplot2)}
}
\value{
NULL (invisibly prints plot)
}
\description{
Print plot silently
}
\keyword{internal}
