% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ssd_hp_ave}
\alias{ssd_hp_ave}
\title{Calculate hazard percent with confidence intervals}
\usage{
ssd_hp_ave(x, conc, nboot)
}
\arguments{
\item{x}{A fitdists object from ssd_fit_bcanz()}

\item{conc}{Numeric concentration value}

\item{nboot}{Integer number of bootstrap samples for confidence intervals}
}
\value{
Data frame with columns: dist, est, se, lcl, ucl, wt
}
\description{
Calculate hazard percent with confidence intervals
}
\keyword{internal}
