% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ui.R
\name{table_lang}
\alias{table_lang}
\title{Language strings for reactable tables}
\usage{
table_lang(lang = "english")
}
\arguments{
\item{lang}{Character string language: "english", "french" or "spanish".}
}
\value{
A \code{\link[reactable:reactableLang]{reactable::reactableLang()}} object.
}
\description{
Language strings for reactable tables
}
\keyword{internal}
