% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tr}
\alias{tr}
\title{Get translation value by ID}
\usage{
tr(id, trans)
}
\arguments{
\item{id}{Character string translation identifier}

\item{trans}{Data frame with columns 'id' and 'trans'}
}
\value{
Character vector of translation text(s)
}
\description{
Get translation value by ID
}
\keyword{internal}
