% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ui.R
\name{ui_dashbox}
\alias{ui_dashbox}
\title{Create a dashed border info box}
\usage{
ui_dashbox(x)
}
\arguments{
\item{x}{UI content to display inside the box}
}
\value{
div element with dashed border styling and centered muted text
}
\description{
Create a dashed border info box
}
\keyword{internal}
