% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ui.R
\name{ui_waiter}
\alias{ui_waiter}
\title{Create a waiter loading screen}
\usage{
ui_waiter(id, ns)
}
\arguments{
\item{id}{Character string output ID to attach waiter to}

\item{ns}{Shiny namespace function}
}
\value{
A Waiter object configured for the namespaced output
}
\description{
Create a waiter loading screen
}
\keyword{internal}
