% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{units}
\alias{units}
\title{Get available concentration units}
\usage{
units()
}
\value{
Character vector of unit strings with Unicode symbols
}
\description{
Get available concentration units
}
\keyword{internal}
