% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ui.R
\name{welcome_card}
\alias{welcome_card}
\title{Create the welcome card}
\usage{
welcome_card(guide_id)
}
\arguments{
\item{guide_id}{Character string namespaced input ID of the link to the
User Guide.}
}
\value{
A card.
}
\description{
The app's purpose and its steps, shown on the Data step until data are
loaded.
}
\keyword{internal}
