% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/about.R
\name{about_page}
\alias{about_page}
\title{Create the About page}
\usage{
about_page(html)
}
\arguments{
\item{html}{Character string rendered About HTML for the current language.}
}
\value{
The page's content.
}
\description{
Create the About page
}
\keyword{internal}
