% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{remove_empty_rows}
\alias{remove_empty_rows}
\title{Remove completely empty rows}
\usage{
remove_empty_rows(data)
}
\arguments{
\item{data}{A data frame or tibble}
}
\value{
A tibble with empty rows removed
}
\description{
Remove completely empty rows
}
\keyword{internal}
