% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-app.R
\name{run_app}
\alias{run_app}
\alias{run_ssdtools_app}
\title{Run Shiny Application}
\usage{
run_app()

run_ssdtools_app()
}
\description{
Run Shiny Application
}
\section{Functions}{
\itemize{
\item \code{run_ssdtools_app}: 
}}

